test(bibleview): split bridge ios tests by owner
Why:
- Adversarial review found the BridgeIOS migration still placed BibleView-owned lifecycle and emission tests in the higher BibleUI package target.
- The migration plan requires tests to land in the lowest package target that owns the behavior after imports are minimized.

What Changed:
- Moved lower-level BibleBridge and WebViewCoordinator iOS tests into BibleViewTests.
- Kept BibleUI reader modal-key routing coverage in BibleUITests.
- Split bridge recording helpers so each package target owns only the helper surface it uses.
- Updated the migration plan to document the BridgeIOS ownership split.

Validation:
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleViewTests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-bridge-ios-bibleview-target CODE_SIGNING_ALLOWED=NO -only-testing:BibleViewTests/BridgeIOSTests
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-bridge-ios-bibleui-target CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/BridgeIOSTests
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleViewTests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-bridge-ios-bibleview-full CODE_SIGNING_ALLOWED=NO
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-bridge-ios-bibleui-full-split CODE_SIGNING_ALLOWED=NO
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild build-for-testing -project AndBible.xcodeproj -scheme AndBible -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-bridge-ios-app-build-split CODE_SIGNING_ALLOWED=NO -skip-testing:AndBibleUITests
- swift package describe --type json
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 -m unittest discover -s scripts -p 'test_*.py'
- git diff --check
- git diff --cached --check
- gitnexus detect_changes scope=staged

Impact:
- Test-only package placement correction with no production behavior change.
- Improves migration fidelity by removing unnecessary BibleUI dependency from BibleView bridge contract tests.
